backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# LangChain and AI
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.6.1

# Vector Database
faiss-cpu==1.7.4
# chromadb==0.4.18

# Document Processing
pypdf==6.0.0
python-docx==1.1.0
docx2txt==0.9
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment and Config
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# HTTP Client
httpx==0.25.2
requests==2.31.0

# Utilities
numpy==1.24.4
pandas==2.1.4
openpyxl==3.1.2
xlrd==2.0.1
tiktoken==0.5.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Monitoring and Logging
prometheus-client==0.19.0
structlog==23.2.0